DS-252 fixes and improvements to page.
Moved the diegram designer CSS into the head as it breaks validation to
put it in the footer.
Added a fix to ensure the diagram popup opens above the page header.
Removed the self closing /'s on BRs and INPUTSs.
Moved the copyright notice HTML comment because it's so long it cuases a
validation error that the meta charset is outside the first 1024 bytes.
Swapped the H2 content to a P with H2 styling as it isn't actually a
title, we just want larger text.
Removed the theme selector and forced light theme, this is a temporary
quick fix for the diagram designer not working with Blinky styles.